added JAVA.md file, divide logic
