+++ README.md
@@ -1,4 +1,6 @@
 # Labs
 
 ## Introduction
-This is the solution of the Labs of the DevOps, please go to special branches to see the solution.
+
+This is the solution of the Labs of the DevOps,
+please go to special branches to see the solution.
