+++ app_java/JAVA.md
@@ -1,0 +1,9 @@
+# Best practices that I used in Java App
+
+- Spring-boot framework that now is a standard of Web App on Java
+- Divide logic into different packages (like controller, view and etc.)
+- Use Lombok that also a good practices in Java code.
+It is simplified classes and automate some things like making
+constructs, getters etc.
+- Use application.properties file for some configs (like server port).
+- Use maven for building and running the project.
