+++ app_java/README.md
@@ -1,23 +1,35 @@
 # Moscow time checker
 
 ## 💾 Information
+
 Sometimes you want to know time in Moscow, but you cant.\
-Now you can easily run this app, go to http://localhost:8080 and see Moscow time!
+Now you can easily run this app, go to <http://localhost:8080> and see Moscow time!
 
 ## 📝 How to run
+
 ### Main dependencies
+
 - Java - 15 or higher
 - Maven - 3.6.3 or higher
+
 ### Building
+
 Go to ``app_java`` directory\
 For building the project, write this command in terminal into the ``app_java`` directory.
-```
+
+``` java
 mvn clean install
 ```
+
 ### Run in one command
-You can also run this app in 1 command, write this command in terminal into the ``app_java`` directory.
-```
+
+You can also run this app in 1 command, write this command
+in terminal into the ``app_java`` directory.
+
+``` java
 mvn spring-boot:run
 ```
+
 ## 📪 Contacts
+
 If you have any questions, please write me in Telegram: `@inforest`
